Windows support through MinGW
